ospf: validate BitStringLength in BIER encapsulation sub-TLV
Bsl::try_from(encap.bs_len).unwrap() will panic when a received LSA
contains an unrecognized BitStringLength value (outside 1..=7). Move
the validation into BierStlv::decode() so that encapsulation sub-TLVs
with unrecognized values are ignored during parsing, and store the
validated Bsl in BierEncapSubStlv instead of a raw u8.

+++ holo-ospf/src/northbound/state.rs
@@ -2317,7 +2317,7 @@ impl<'a> YangList<'a, Instance<Ospfv3>> for ospf::areas::area::database::area_sc
         Self {
             max_si: Some(bier_encap.max_si),
             id: Some(bier_encap.id.clone().get()),
-            bs_len: Some(bier_encap.bs_len)
+            bs_len: Some(u8::from(bier_encap.bs_len))
         }
     }
 }
